Use consistent nameing for MethodParameter
